requirements.txt
# configuration
#oslo.config==3.2.0    # configuration from openstack oslo
oslo.config==4.8.0

# security
passlib==1.6.5
pycrypto==2.6.1
rsa==3.4.2

# time
pytz==2015.7
tzlocal==1.2
iso8601==0.1.11

# database
PyMySQL==0.7.0
SQLAlchemy==1.2.2
Alembic==0.8.4         # sqlalchemy migration

cryptography==2.3.1

# openstack
python-keystoneclient==2.1.1
python-glanceclient==1.2.0
python-novaclient==2.33.0
python-neutronclient==4.1.1
python-cinderclient==1.5.0
python-swiftclient==2.7.0
futures==3.1.1

# web
Flask==0.10.1
Flask-Script==2.0.5
Flask-Bootstrap==3.3.5.7
Flask-Moment==0.5.1
Flask-SQLAlchemy==2.1
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-WTF==0.12
Flask-Babel==0.9
WTForms==2.1
Flask-Login==0.3.2
Flask-SSLify==0.1.5
Flask-httpauth==3.1.1

# TODO Add Flask-Security to support token authentication and complicated security features
# TODO Need a resource management library like Flask-RBAC

# messaging(need install before celery, celery need amqp==2.1.4 kombu==4.0.2, oslo.messaging need amqp==2.1.3 kombu==4.0.1)
oslo.messaging==5.27.0

# task queue with rabbitmq
#Celery==3.1.20
Celery==4.0.2

# fix dependency issue for kombu, amqp, oslo_log
kombu==4.0.2
amqp==2.1.4
oslo.log==3.28.1

#Celery==3.1.20
#kombu==3.0.33
#amqp==1.4.9

# excel uploads
xlrd==0.9.4

# fix dependency issue for Flask-Excel(0.0.4),
# Flask-Excel(0.0.4) default install pyexcel-webio >= 0.0.5,
# but pyexcel-webio==0.1.2 is not compatible with Flask-Excel(0.0.4)
# related issue #566

pyexcel-webio==0.0.11
pyexcel-io==0.1.0       # not 0.2.0
pyexcel-xls==0.1.0      # not 0.2.0
pyexcel-xlsx==0.1.0     # not 0.2.0
pyexcel==0.2.0          # not 0.2.2
Flask-Excel==0.0.4

# auth
Flask-Principal==0.4.0

# voluptuous data validation
voluptuous==0.8.9

# intervaltree
intervaltree==2.1.0

# zookeeper client
kazoo==2.2.1

# coroutine for gateway and websocket
gevent==1.1.2

# websocket
redis==2.10.5
Flask-uWSGI-WebSocket==0.6.0

#influxdb
influxdb==4.1.1

# hybrid computation
configparser==3.5.0

# host power management
pynacl==1.2.1
bcrypt==3.1.4
paramiko==2.4.2

